feat: Content에서 알람 버튼을 눌렀을때에 알람 시간이 저장되고 알람 시간이 띄워진다.

+++ app/src/main/java/com/kt_study/todo_alarm/db/ContentEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "table_content")
 data class ContentEntity(
     @PrimaryKey
-    val id: Long = 0L,
+    val id: Int = 0,
     val categoryId: Long = 0L,
     var toDo: String = "",
     var hour: Int = 0,
